Correcao/requirements.txt
python-dotenv 
mysql-connector-python
pandas
jinja2
